% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_filter.R
\name{filter_cover}
\alias{filter_cover}
\title{Function to check coverage of gene-pathway mapping}
\usage{
filter_cover(X, pathway, filt_cov)
}
\arguments{
\item{X}{matrix of data.}

\item{pathway}{list of pathways to analysis}

\item{filt_cov}{numeric vector of length 1. Minimum \% of pathway coverage after gene mapping. By default is 0 (no filtration), the 0.65 will indicate that only pathway with coverage of genes larger than 65\% will be taken.}
}
\value{
A list of filtered pathways.
}
\description{
The function...
}
