% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmm_support.R
\name{find_last_consecutive_true}
\alias{find_last_consecutive_true}
\title{Function to find last consecutive true}
\usage{
find_last_consecutive_true(row)
}
\arguments{
\item{row}{vector of data}
}
\value{
A vector of ranked genes.
}
\description{
The function...
}
