% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2path.R
\name{gene2path}
\alias{gene2path}
\title{Function to transform gene level information to pathway level by single-sample approaches}
\usage{
gene2path(X, pathway, method = "ssGSEA")
}
\arguments{
\item{X}{matrix of data.}

\item{pathway}{list of pathways to analsysis}

\item{method}{methods for single-sample pathway enrichment analysis.}
}
\value{
Function returns...
}
\description{
Function reduce genes to pathways.
}
\examples{
\dontrun{


}

}
\seealso{
\code{\link{gaussian_mixture_vector}}, \code{\link{EM_iter}}
}
