% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMdecomp.R
\name{GMMdecomp}
\alias{GMMdecomp}
\title{Function for GMM decomposition of pathways}
\usage{
GMMdecomp(X, K = 10, multiply = TRUE, parallel = FALSE)
}
\arguments{
\item{X}{matrix of data enrichment scores rows pathways columns samples.}

\item{K}{maximum number of components for GMM decomposition (default: 10).}

\item{multiply}{logical, scaling values by 10 before fitting GMM (default: TRUE).}

\item{parallel}{logical, lunching parallel computing (default: FALSE).}
}
\value{
Function returns...
}
\description{
Function reduce genes to pathways.
}
