% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMthr.R
\name{GMMthr}
\alias{GMMthr}
\title{Function for selecting thresholds ofor pathway activity}
\usage{
GMMthr(gmms)
}
\arguments{
\item{gmms}{output from GMMdecomp function.}
}
\value{
Function returns...
}
\description{
Function reduce genes to pathways.
}
