% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JASMINE_support_func.R
\name{LikelihoodCalculation}
\alias{LikelihoodCalculation}
\title{Function to calculated Likelihood in JASMINE pathway enrichment}
\source{
\url{https://github.com/NNoureen/JASMINE}
}
\usage{
LikelihoodCalculation(X, g_vec)
}
\arguments{
\item{X}{matrix of data.}

\item{g_vec}{vector of gene names from pathway.}
}
\value{
A vector of likelihood.
}
\description{
The function...
}
