% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_methods.R
\name{Mean_path}
\alias{Mean_path}
\title{Function to perform Mean pathway enrichment}
\usage{
Mean_path(pathway, X)
}
\arguments{
\item{pathway}{list of pathways to analysis}

\item{X}{matrix of data.}
}
\value{
A data.frame with pathways in rows and samples in columns.
}
\description{
The function...
}
