% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover_functions.R
\name{min_max_filter}
\alias{min_max_filter}
\title{Function to remove pathways from analysis by their size.}
\usage{
min_max_filter(pathway, filt_min, filt_max)
}
\arguments{
\item{pathway}{list of pathways to analysis}

\item{filt_min}{numeric vector of length 1. Minimum size of the resulting pathway after gene identifier mapping. By default, the minimum size is 15.}

\item{filt_max}{numeric vector of length 1. Maximum size of the resulting pathway after gene identifier mapping. By default, the minimum size is 500.}
}
\value{
A list of filtered pathways.
}
\description{
The function...
}
