% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{Norm_JASMINE}
\alias{Norm_JASMINE}
\title{Function to normalize JASMINE score}
\source{
\url{https://github.com/NNoureen/JASMINE}
}
\usage{
Norm_JASMINE(data)
}
\arguments{
\item{JAS_Scores}{vector of of data.}
}
\value{
A vector of normalized JASMINE score.
}
\description{
The function...
}
