% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JASMINE_support_func.R
\name{NormalizationJAS}
\alias{NormalizationJAS}
\title{Function to normalize JASMINE score}
\source{
\url{https://github.com/NNoureen/JASMINE}
}
\usage{
NormalizationJAS(JAS_Scores)
}
\arguments{
\item{JAS_Scores}{vector of of data.}
}
\value{
A vector of normalized JASMINE score.
}
\description{
The function...
}
