% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{ORCalculation}
\alias{ORCalculation}
\title{Function to calculated Odds Ratio in JASMINE pathway enrichment}
\source{
\url{https://github.com/NNoureen/JASMINE}
}
\usage{
ORCalculation(X, g_vec)
}
\arguments{
\item{X}{matrix of data.}

\item{g_vec}{vector of gene names from pathway.}
}
\value{
A vector of odds ratios.
}
\description{
The function...
}
