% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{Path_df}
\alias{Path_df}
\title{Function to calculate JASMINE score pathway enrichment}
\usage{
Path_df(g_vec, X)
}
\arguments{
\item{g_vec}{vector of gene names from pathway.}

\item{X}{matrix of data.}
}
\value{
data frame
}
\description{
The function...
}
