% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_methods.R
\name{Path_JASMINE}
\alias{Path_JASMINE}
\title{Function to perform JASMINE pathway enrichment}
\source{
\url{https://github.com/NNoureen/JASMINE}
}
\usage{
Path_JASMINE(X, pathway, type = "oddsratio")
}
\arguments{
\item{X}{matrix of data.}

\item{pathway}{list of pathways to analysis}
}
\value{
A data.frame with pathways in rows and samples in columns.
}
\description{
The function...
}
