% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_methods.R
\name{Path_ssGSEA}
\alias{Path_ssGSEA}
\title{Function to perform ssGSEA pathway enrichment}
\usage{
Path_ssGSEA(X, pathway)
}
\arguments{
\item{X}{matrix of data.}

\item{pathway}{list of pathways to analysis}
}
\value{
A data.frame with pathways in rows and samples in columns.
}
\description{
The function...
}
