% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_methods.R
\name{pathBINA}
\alias{pathBINA}
\title{Function to perform BINA pathway enrichment}
\usage{
pathBINA(X, pathway)
}
\arguments{
\item{X}{matrix or data.frame of data (rows: genes/features, columns: samples).}

\item{pathway}{list of pathways to analysis.}
}
\value{
A data.frame with pathways in rows and samples in columns.
}
\description{
The function...
}
