% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_methods.R
\name{pathCERNO}
\alias{pathCERNO}
\title{Wrapper to CERNO pathway enrichment}
\usage{
pathCERNO(X, pathway)
}
\arguments{
\item{X}{A numeric matrix or data.frame with genes/features as rows and samples as columns.
Row names (gene identifiers) must be provided.}

\item{pathway}{A named list of pathways, where each element is a character vector of gene identifiers.}
}
\value{
A data.frame with pathways as rows and samples as columns containing pathway activity scores.
}
\description{
Implements the CERNO pathway enrichment test, which calculates pathway activity
  using a non-parametric approach based on gene expression ranks. For each gene set, the method evaluates whether
  the genes in the set are enriched toward the top (or bottom) of a ranked gene list for each sample.
  Specifically, it computes an Area Under the Curve (AUC)-like statistic derived from the Mann–Whitney U test.
}
\details{
The AUC score is computed as:
  \deqn{
  \text{AUC}_G = \frac{n_G (n - n_G) + \frac{n_G(n_G + 1)}{2} - R_G}{n_G (n - n_G)}
  }
  where \( n \) is the total number of genes, \( n_G \) is the number of genes in the pathway,
  and \( R_G \) is the sum of their ranks in a given sample.

  Scores closer to 1 indicate that genes in the pathway tend to have higher ranks (greater activity),
  while values closer to 0 suggest lower relative activity.

  This approach enables robust and distribution-free inference of pathway activity, particularly useful
  in transcriptomic analyses. Its extensive implementation was presented in Zyla et al. (2019).
}
\references{
Zyla, J., Marczyk, M., Domaszewska, T., Kaufmann, S. H., Polanska, J., & Weiner III, J. (2019).
Gene set enrichment for reproducible science: comparison of CERNO and eight other algorithms.
*Bioinformatics*, *35*(24), 5146–5154.
\doi{10.1093/bioinformatics/btz447}
}
