% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_methods.R
\name{pathssGSEA}
\alias{pathssGSEA}
\title{Function to perform ssGSEA pathway enrichment}
\source{
\url{https://github.com/rcastelo/GSVA}
}
\usage{
pathssGSEA(X, pathway)
}
\arguments{
\item{X}{matrix of data.}

\item{pathway}{list of pathways to analysis.}
}
\value{
A data.frame with pathways in rows and samples in columns.
}
\description{
The function...
}
\references{
Barbie, D.A. et al. Systematic RNA interference reveals that
oncogenic KRAS-driven cancers require TBK1.
*Nature*, 462(5):108-112, 2009.
}
