% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways}
\alias{pathways}
\title{Exemplary data of pathways}
\format{
A list with 45 elements which represent pathways. Each element contain vector of gene names (characters) of each pathway.
Names represent CellMarker pathway ID for which information are given in pathways_info file.
}
\source{
\url{https://cran.r-project.org/web/packages/tmod/index.html}
}
\usage{
data(pathways)
}
\description{
File containing 45 pathways form CellMarker database.
}
\keyword{datasets}
