% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways_info}
\alias{pathways_info}
\title{Pathay description for exemplary data}
\format{
A data frame with 45 rows and 3 columns variables:
\describe{
  \item{ID}{ID pathways ID which math to names of pathway data.}
  \item{Title}{Title pathways title}
  \item{DataBase}{DataBase information about database for obtaining the pathways.}
}
}
\source{
\url{https://cran.r-project.org/web/packages/tmod/index.html}
}
\usage{
data(pathways_info)
}
\description{
File containing 45 pathways form CellMarker database.
}
\keyword{datasets}
