% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_methods.R
\name{pathZScore}
\alias{pathZScore}
\title{Function to perform Z-score integration pathway enrichment}
\usage{
pathZScore(X, pathway)
}
\arguments{
\item{X}{matrix of data.}

\item{pathway}{list of pathways to analysis.}
}
\value{
A data.frame with pathways in rows and samples in columns.
}
\description{
The function...
}
