% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{RankCalculation}
\alias{RankCalculation}
\title{Function to rank data on JASMINE pathway enrichment}
\source{
\url{https://github.com/NNoureen/JASMINE}
}
\usage{
RankCalculation(X, g_vec)
}
\arguments{
\item{X}{matrix of data.}

\item{g_vec}{vector of gene names from pathway.}
}
\value{
A vector of ranked genes.
}
\description{
The function...
}
