% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{rowAUC}
\alias{rowAUC}
\title{Function to calculate AUC fo CERNO method}
\usage{
rowAUC(X, df)
}
\arguments{
\item{X}{matrix of data.}

\item{df}{data.frame from CERNO function}
}
\value{
data frame
}
\description{
The function...
}
