% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_methods.R
\name{ssGSEA}
\alias{ssGSEA}
\title{Function to performe ssGSEA pathway enrichment}
\usage{
ssGSEA(X, pathway)
}
\arguments{
\item{X}{matrix of data.}

\item{pathway}{list of pathways to analsysis}
}
\value{
A data.frame with pathways in rows and samples in columns.
}
\description{
The function...
}
