% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thr_AUCell.R
\name{thr_AUCell}
\alias{thr_AUCell}
\title{Function for threshold pathway activity scores by AUCell approach.}
\source{
\url{https://www.bioconductor.org/packages/release/bioc/html/AUCell.html}
}
\usage{
thr_AUCell(df_path)
}
\arguments{
\item{df_path}{matrix or data.frame of pathway activity scores (rows: path, columns: samples).}
}
\value{
vector of thresholds for AUCell method.
}
\description{
Function search for thresholds which indicates samples with relative significant pathway activity.
The process is performed based on AUCell methods
}
\references{
Aibar S, Bravo Gonzalez-Blas C, Moerman T, Huynh-Thu V, Imrichova H, Hulselmans G, Rambow F, Marine J, Geurts P, Aerts J, van den Oord J, Kalender Atak Z, Wouters J, Aerts S (2017). “SCENIC: Single-Cell Regulatory Network Inference And Clustering.” Nature Methods, 14, 1083-1086. doi:10.1038/nmeth.4463.
}
