% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thr_GMM.R
\name{thr_GMM}
\alias{thr_GMM}
\title{Function for selecting thresholds for pathway activity based on GMM decoposition.}
\usage{
thr_GMM(gmms)
}
\arguments{
\item{gmms}{output from GMMdecomp function or list of outputs from dpGMM package.
Each first list level should refer to the decomposition of one pathway.
Second level of list should include variable threshold (vector of thresholds) and model (matrix with model parameters).}
}
\value{
List of thresholds for each investigated pathway based on GMM decomposition.
}
\description{
Function search for thresholds which indicates samples with relative significant pathway activity.
The process is performed for GMM decomposition output.
}
