% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thrGMM.R
\name{thrGMM}
\alias{thrGMM}
\title{Function for GMM threshold}
\usage{
thrGMM(X, K = 10, parallel = F)
}
\arguments{
\item{X}{matrix of data enrichment scores rows pathways columns samples.}

\item{K}{maximum number of components for GMM decomposition}

\item{parallel}{default FALSE. Lunching parallel computing.}
}
\value{
Function returns...
}
\description{
Function reduce genes to pathways.
}
