% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{Calc_AUC}
\alias{Calc_AUC}
\title{Function to calculate AUC for CERNO pathway enrichment}
\usage{
Calc_AUC(X, df)
}
\arguments{
\item{X}{matrix of data.}

\item{df}{data.frame of ranked genes in pathway.}
}
\value{
a vector of CERNO AUC calculation.
}
\description{
The function...
}
