% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{extract_pathway}
\alias{extract_pathway}
\title{Function to extract matrix of genes in pathway}
\usage{
extract_pathway(X, g_vec)
}
\arguments{
\item{X}{matrix of data.}

\item{g_vec}{vector of gene names from pathway.}
}
\value{
A data.frame with pathway genes in rows and samples in columns.
}
\description{
The function...
}
