% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_filter.R
\name{filter_minmax}
\alias{filter_minmax}
\title{Filter Pathways by Size}
\usage{
filter_minmax(pathway, filt_min, filt_max)
}
\arguments{
\item{pathway}{A named list of pathways, where each element is a character vector of gene identifiers.}

\item{filt_min}{Integer specifying the minimum size (number of genes) a pathway must have after mapping to \code{X}.
Default is 15.}

\item{filt_max}{Integer specifying the maximum size (number of genes) a pathway can have after mapping to \code{X}.
Default is 500.}
}
\value{
A list of filtered pathways.
}
\description{
This function removes pathways from a list based on the number of genes they contain. It is typically used
prior to enrichment analysis to exclude pathways that are too small or too large, which may bias the results.
}
