% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMdecomp.R
\name{GMMdecomp}
\alias{GMMdecomp}
\title{Function for GMM decomposition of pathways}
\usage{
GMMdecomp(X, K = 10, multiply = TRUE, IC = "BIC", parallel = FALSE)
}
\arguments{
\item{X}{data.frame of data enrichment scores (rows: pathways, columns: samples).}

\item{K}{Maximum number of components for GMM decomposition (default: 10).}

\item{multiply}{Logical, whether to scale values by 10 before fitting GMM (default: TRUE).}

\item{IC}{Information criterion used to select the number of model components.
Possible methods are "AIC","AICc", "BIC" (default), "ICL-BIC" or "LR".}

\item{parallel}{Logical, whether to use parallel computing (default: FALSE). (Not implemented yet)}
}
\value{
Function returns...
}
\description{
Function reduce genes to pathways.
}
