% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMdecomp.R, R/thrGMM.R
\name{GMMdecomp}
\alias{GMMdecomp}
\title{Function for GMM decomposition of pathways}
\usage{
GMMdecomp(X, K = 10, multiply = T, parallel = F)

GMMdecomp(X, K = 10, multiply = T, parallel = F)
}
\arguments{
\item{X}{matrix of data enrichment scores rows pathways columns samples.}

\item{K}{maximum number of components for GMM decomposition}

\item{parallel}{default FALSE. Lunching parallel computing.}
}
\value{
Function returns...

Function returns...
}
\description{
Function reduce genes to pathways.

Function reduce genes to pathways.
}
