% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{JASMINE}
\alias{JASMINE}
\title{Function to calculate JASMINE score pathway enrichment}
\source{
\url{https://github.com/NNoureen/JASMINE}
}
\usage{
JASMINE(X, g_vec, type = "oddsratio")
}
\arguments{
\item{X}{matrix of data.}

\item{g_vec}{vector of gene names from pathway.}

\item{type}{type of adjustemnt of JASMINE score. By default 'oddsratio", another possible input is "likelihood". Parameter only valid for JASMINE method.}
}
\value{
A vector of normalized JASMINE score.
}
\description{
The function...
}
