% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_gmm_support.R
\name{km_search}
\alias{km_search}
\title{Function to find threshold based on GMM parameters}
\usage{
km_search(params, thrs)
}
\arguments{
\item{params}{matrix of GMM decomposition parameters. Rows represent clusters columns represent GMM parameters with following order: mean, sd, alpha.}

\item{thrs}{vector of thresholds of GMM decomposition.}
}
\value{
a numeric value with threshold.
}
\description{
The function...
}
