% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_methods.R
\name{Path_Cerno}
\alias{Path_Cerno}
\title{Function to perform CERNO pathway enrichment}
\usage{
Path_Cerno(X, pathway)
}
\arguments{
\item{X}{matrix of data.}

\item{pathway}{list of pathways to analysis}
}
\value{
A data.frame with pathways in rows and samples in columns.
}
\description{
The function...
}
