% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_methods.R
\name{pathCERNO}
\alias{pathCERNO}
\title{Function to perform CERNO pathway enrichment}
\usage{
pathCERNO(X, pathway)
}
\arguments{
\item{X}{matrix or data.frame of data (rows: genes/features, columns: samples).}

\item{pathway}{list of pathways to analysis.}
}
\value{
A data.frame with pathways in rows and samples in columns.
}
\description{
The function...
}
\references{
Zyla, J. et al. Gene set enrichment for reproducible science: comparison of CERNO and
eight other algorithms. Bioinformatics, 2019, 35.24: 5146-5154.
}
