% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_methods.R
\name{pathJASMINE}
\alias{pathJASMINE}
\title{Function to perform JASMINE pathway enrichment}
\source{
\url{https://github.com/NNoureen/JASMINE}
}
\usage{
pathJASMINE(X, pathway, type = "oddsratio")
}
\arguments{
\item{X}{X matrix or data.frame of data (rows: genes/features, columns: samples).}

\item{pathway}{list of pathways to analysis.}

\item{type}{type of adjustment of JASMINE score. By default 'oddsratio", another possible input is "likelihood". Parameter only valid for JASMINE method.}
}
\value{
A data.frame with pathways in rows and samples in columns.
}
\description{
The function...
}
\references{
Noureen, N. et al. Signature-scoring methods developed for bulk samples are
not adequate for cancer single-cell RNA sequencing data. Elife, 2022, 11: e71994.
}
