% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{Rank_data}
\alias{Rank_data}
\title{Function to rank data}
\usage{
Rank_data(X)
}
\arguments{
\item{X}{matrix of data.}
}
\value{
A data.frame with ranked genes rows and samples in columns.
}
\description{
The function...
}
