% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{Rank_dropout}
\alias{Rank_dropout}
\title{Function to rank data without zero counts}
\source{
\url{https://github.com/NNoureen/JASMINE}
}
\usage{
Rank_dropout(X, g_vec)
}
\arguments{
\item{X}{matrix of data.}

\item{g_vec}{vector of gene names from pathway.}
}
\value{
A vector of ranked genes.
}
\description{
The function...
}
