% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{scale_minmax}
\alias{scale_minmax}
\title{Function to min-max normalization}
\usage{
scale_minmax(x)
}
\arguments{
\item{x}{vector of of data.}
}
\value{
A vector of normalized data.
}
\description{
The function...
}
