% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{scale_minmax}
\alias{scale_minmax}
\title{Function for Min-Max Normalization}
\usage{
scale_minmax(x)
}
\arguments{
\item{x}{A numeric vector of data to be normalized.}
}
\value{
A numeric vector of the same length as \code{x}.
}
\description{
This function applies min-max normalization to a numeric vector, scaling the values to the range [0, 1].
It is commonly used to standardize feature values before further analysis or scoring.
}
\keyword{internal}
