% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ss_support.R
\name{scale_zscore}
\alias{scale_zscore}
\title{Function to z-score normalization}
\usage{
scale_zscore(X)
}
\arguments{
\item{X}{matrix or vector of of data.}
}
\value{
A matrix or vector of normalized data.
}
\description{
The function...
}
