% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thr_AUCell.R
\name{thr_AUCell}
\alias{thr_AUCell}
\title{Thresholding pathway activity scores using the AUCell approach}
\source{
\url{https://www.bioconductor.org/packages/release/bioc/html/AUCell.html}
}
\usage{
thr_AUCell(df_path)
}
\arguments{
\item{df_path}{A numeric matrix or data.frame of pathway activity scores,
where rows correspond to pathways and columns to samples.}
}
\value{
A vector of threshold values.
  For detailed behavior, refer to the AUCell package documentation.
}
\description{
Applies the AUCell package's methodology to identify thresholds that indicate
samples with relatively high pathway activity. This function determines, for each
pathway, which samples exhibit significant enrichment based on expression rankings.
}
\references{
Aibar, S., Bravo González-Blas, C., Moerman, T., Huynh-Thu, V.A.,
  Imrichová, H., Hulselmans, G., Rambow, F., Marine, J.C., Geurts, P., Aerts, J.,
  van den Oord, J., Kalender Atak, Z., Wouters, J., & Aerts, S. (2017).
  SCENIC: Single-Cell Regulatory Network Inference And Clustering. *Nature Methods*, 14, 1083–1086.
  \doi{10.1038/nmeth.4463}
}
